Moving UnspacedList to a parse_common.py
